RestSend, Results: adhere to property best practice
Python property semantics should be symmetric.  That is, you should be able to:

x = obj.my_property
obj.my_property = x

And expect the same value type in both cases.

For RestSend and Results we were violating these semantics.

Resolved this by:

1. Adding methods to replace the property setters:

- add_changed()
- add_failed()
- add_response()
- add_result()

2. Removing the setters.

3. Adding ## See also sections to docstrings for these properties and methods.

Now, the intent is clear for users of these classes, and linters will not complain.

Other changes

1. Clean up all docstrings in both classes
2. RestSend: methods commit_normal_mode and common_check_mode should be private.  Fixed.